+++ Verba/VerbaApp.swift
@@ -16,7 +16,11 @@ class AppDelegate: NSObject, NSApplicationDelegate {
 struct VerbaApp: App {
     @NSApplicationDelegateAdaptor(AppDelegate.self) var appDelegate
     @StateObject private var appState = AppState()
+    #if DEBUG
+    private let updaterController = SPUStandardUpdaterController(startingUpdater: false, updaterDelegate: nil, userDriverDelegate: nil)
+    #else
     private let updaterController = SPUStandardUpdaterController(startingUpdater: true, updaterDelegate: nil, userDriverDelegate: nil)
+    #endif
 
     var body: some Scene {
         MenuBarExtra {